Moved "controller" files to their own directory and added rewrite rules to server config so that they are accessed (an example nginx config has been added to config/). And renamed some files. Fixed most (all?) hrefs to work with the new directory structure.
Also, using rewrite rules, I made it so that the id for any pages
accessing a single DB resource appear in a pretty way in the URI. E.g.
for the Player Profile page, the URI looks like /player/21. In the
future, I expect to just add extra arguments after a '?'.

Rewrote some files that weren't proper MVC and fixed all bugs that
occurred. Except for Calendar, the site is free of the most obvious
bugs.

Removed image files from version control.